fix(syncnode): wave 2 hardening — 6 remaining audit findings + dead-route discovery
Closes wave-2 stability/performance findings from the audit. 13 syncnode
packages green under -race -short with 453 tests (up from 450 in
wave 1).

S4 — lock-order documentation + release-before-call
(master/cluster.go + sync_node_task.go):
  checkSyncNodeHeartbeat now snapshots IsActive under sn.RLock(release)
  before AND after checkLiveness() so the read race is gone, and
  handleNodeDeath is invoked AFTER all SyncNode locks are released.
  Canonical lock-order header comment landed above
  checkSyncNodeHeartbeat:
    1. Cluster.syncNodes  →  2. SyncNode.RWMutex
    →  3. SyncDispatcher.mu  →  4. SyncFailover.mu
  Code paths touching multiple levels MUST acquire top-down.

S5 — split-brain on syncnode reconnect (master/sync_node_task.go):
  handleSyncNodeHeartbeatResp captures wasActive under the existing
  sn.Lock; on the false→true edge calls a new private
  cancelOrphanedTasksOnReconnect(addr) which iterates the dispatcher's
  AllOwnerships() and queues OpSyncNodeCancelTask for every taskID
  whose owner != addr (i.e. master had re-homed it during the outage).
  Syncnode's TaskHandler picks up the cancel and stops the orphaned
  local run, preventing double-execution. Fan-out is O(N) only on the
  rare reconnect transition; happy-path heartbeats stay O(1).

SEC5 — quota float64 truncation
(syncnode/ratelimit/* + master/sync_quota.go):
  Bucket.SetLimit / Bucket.Mbps / Registry.SetRuleLimit /
  Registry.SetBackendLimit / Registry.RuleSnapshot / Snapshot signatures
  now float64. NewBucket(int) and NewRegistry(int) UNCHANGED to avoid
  churn at the static node-level constructor. applyQuotaReply +
  applyHeartbeatReply forward reply.RuleQuotas / BackendQuotas
  untruncated. Cluster cap now strictly preserved (was off by up to
  N-1 MBps due to per-node floor-divide).

P1 — Snapshot cache (syncnode/snapshot.go + server.go):
  Moved the synchronous BoltDB scans (computeRecentFailureRate +
  advertiseRules) off the heartbeat hot path into a snapshotCache
  struct with atomic.Value / atomic.Pointer reads + a background
  ticker that refreshes every snapshotCacheRefresh=30s. Snapshot()
  now O(1) on the hot path. Immediate-seed on startSnapshotCacheLoop
  closes the 0→30s startup gap.

P2 — sync_quota.Compute in-place map reuse (master/sync_quota.go):
  Added gen uint64 + perNodeGen map[string]uint64. Compute() bumps
  the generation, reuses existing NodeQuotas entries (clearing inner
  Rules/Backends maps via delete-loop instead of re-allocating), then
  sweeps stale entries (gen != current). For 100 nodes × 50 rules
  the prior 5000+ entries/tick allocation churn is gone. Verified
  via reflect.ValueOf(map).Pointer() identity test across two calls.

P3 — listSyncNodes O(N²) → O(N)
(master/sync_dispatcher.go + api_service.go):
  New SyncDispatcher.LoadScoreAll() walks rangeSyncNodes ONCE with a
  single now() snapshot, returning map[addr]score. listSyncNodes
  calls it once then map-lookups per node. For 1000 nodes the prior
  1M operations/request drops to 1k.

P4 + S7 — quota piggy-back + missing /syncNode/response route
(master/api_service.go + http_server.go + sdk/master/api_node.go +
syncnode/master_client.go):
  Discovered during wave 2: master/http_server.go NEVER WIRED the
  /syncNode/response route. handleSyncNodeTaskResponse was dead code;
  the syncnode → master heartbeat push (and terminal pushback) got 404
  silently. Master never saw runtime stats — RunningTasks /
  BoltDBHealthy / etc. on master/SyncNode struct stayed zero, making
  the load-score formula degenerate.

  FIX P4 combined the route fix with quota piggy-back:
    - new master.handleSyncNodeTaskResponse HTTP handler wired at
      proto.GetSyncNodeTaskResponse: drives cluster state update,
      then replies with newSuccessHTTPReply(*SyncNodeQuotaReply).
    - SDK NodeAPI.ResponseSyncNodeTask now returns
      (*proto.SyncNodeQuotaReply, error) via mc.requestWith — one
      round-trip per heartbeat instead of two.
    - syncnode/master_client.sendHeartbeat applies the inline reply
      via applyQuotaReply; nil reply (older master build) tolerated.
    - GET /syncNode/getQuota route removed; SDK GetSyncNodeQuota
      marked DEPRECATED (kept for one release as migration window —
      callers on old syncnode builds see a 404 surfacing "upgrade
      required" rather than silent fall-through).

  For 100 syncnodes this reclaims ~600 RPS on master + 100 TLS
  handshakes per 10s tick.

Tests added in this commit: 15
  - master/sync_node_task_test.go: 3 reconnect-detection tests
  - master/sync_quota_test.go: 3 map-reuse + sweep tests
  - master/sync_dispatcher_test.go: 3 LoadScoreAll tests
  - syncnode/snapshot_test.go: 3 cache-pattern tests
  - syncnode/ratelimit/*_test.go: cosmetic float64 signature updates

Files touched: 18 — see git diff.

Pre-existing macOS rocksdb cgo issue still blocks `go test ./master/...`
locally; Linux CI runs the master suite.

+++ syncnode/ratelimit/limiter.go
@@ -51,20 +51,32 @@ type Limiter interface {
 // "unlimited" — WaitN returns nil immediately without consulting the
 // underlying rate.Limiter.
 //
+// The configured rate is stored as float64 because cluster-wide caps
+// distributed across N nodes (sync_quota.Compute) produce fractional MB/s
+// (e.g. 400 / 7 ≈ 57.14). Truncating to int under-enforced the total cap
+// by up to N-1 MB/s (SEC5). NewBucket accepts int for the static
+// constructor convenience used by server.go / executor.go; SetLimit takes
+// float64 since the dynamic-update path receives fractional values from
+// master.
+//
 // Bucket is safe for concurrent use.
 type Bucket struct {
 	mu   sync.Mutex
 	rl   *rate.Limiter // nil when unlimited
-	mbps int           // configured Mbps, kept for SetLimit / diagnostics
+	mbps float64       // configured Mbps, kept for SetLimit / diagnostics
 }
 
 // NewBucket constructs a token bucket capped at mbps megabytes per second.
 // mbps <= 0 is treated as unlimited; the returned Bucket is still valid and
 // its WaitN is a no-op. Burst defaults to max(1 second of bandwidth,
 // 4 MiB) so freshly minted buckets don't stall the first Read of a stream.
+//
+// Accepts int for backward compatibility with the static node-level
+// constructor at server boot. Dynamic / fractional retuning goes through
+// SetLimit which takes float64.
 func NewBucket(mbps int) *Bucket {
 	b := &Bucket{}
-	b.setLimitLocked(mbps)
+	b.setLimitLocked(float64(mbps))
 	return b
 }
 
@@ -73,19 +85,23 @@ func NewBucket(mbps int) *Bucket {
 // immediately). Safe for concurrent use; callers in flight on the old
 // limiter complete their wait against the old rate, future callers see the
 // new one.
-func (b *Bucket) SetLimit(mbps int) {
+//
+// Takes float64 because cross-node quotas (per-rule / per-backend) come
+// from master's equal-division of cluster caps and are commonly
+// fractional. See SEC5 in the design notes.
+func (b *Bucket) SetLimit(mbps float64) {
 	b.mu.Lock()
 	defer b.mu.Unlock()
 	b.setLimitLocked(mbps)
 }
 
-func (b *Bucket) setLimitLocked(mbps int) {
+func (b *Bucket) setLimitLocked(mbps float64) {
 	b.mbps = mbps
 	if mbps <= 0 {
 		b.rl = nil
 		return
 	}
-	bytesPerSec := float64(mbps) * 1024 * 1024
+	bytesPerSec := mbps * 1024 * 1024
 	burst := int(bytesPerSec)
 	if burst < minBurstBytes {
 		burst = minBurstBytes
@@ -95,7 +111,7 @@ func (b *Bucket) setLimitLocked(mbps int) {
 
 // Mbps returns the configured rate (0 means unlimited). Exposed for
 // diagnostics / metrics; not used in the hot path.
-func (b *Bucket) Mbps() int {
+func (b *Bucket) Mbps() float64 {
 	b.mu.Lock()
 	defer b.mu.Unlock()
 	return b.mbps
